🚚 Rename: article 에서 board 로 변경
